Fix for warning: C4267 - Data loss due to downcasting
warning C4267: 'return': conversion from 'size_t' to 'int', possible loss of data

Fix: Converted to use size_t where possible. See: ISO/IEC 9899:201x for more info, but this should improve inter-library compatibility.

+++ include/Audijo/AsioApi.hpp
@@ -41,7 +41,7 @@ namespace Audijo
 		
 		const std::vector<DeviceInfo<Asio>>& Devices(bool reload = false);
 		const DeviceInfo<>& Device(int id) const override { for (auto& i : m_Devices) if (i.id == id) return i; };
-		int DeviceCount() const override { return m_Devices.size(); };
+		size_t DeviceCount() const override { return m_Devices.size(); };
 		const DeviceInfo<Asio>& ApiDevice(int id) const { for (auto& i : m_Devices) if (i.id == id) return i; };
 
 		Error Open(const StreamParameters& settings = StreamParameters{}) override;
